+++ packages/transaction-pay-controller/src/strategy/bridge/bridge-quotes.ts
@@ -39,7 +39,6 @@ const SLIPPAGE_DEFAULT = 0.005;
 
 const FEATURE_ID_BY_TRANSACTION_TYPE = new Map<TransactionType, FeatureId>([
   [TransactionType.perpsDeposit, FeatureId.PERPS],
-  [TransactionType.perpsDepositAndOrder, FeatureId.PERPS],
 ]);
 
 const log = createModuleLogger(projectLogger, 'bridge-strategy');