+++ packages/transaction-pay-controller/src/utils/source-amounts.ts
@@ -232,8 +232,7 @@ function isQuoteAlwaysRequired(
     isHyperliquidDeposit &&
     (strategy === TransactionPayStrategy.Relay ||
       (strategy === TransactionPayStrategy.Across &&
-        (parentTransactionType === TransactionType.perpsDeposit ||
-          parentTransactionType === TransactionType.perpsDepositAndOrder)))
+        parentTransactionType === TransactionType.perpsDeposit))
   );
 }